requirements.txt
asgiref==3.2.10
certifi==2020.6.20
dj-database-url==0.5.0
Django==3.1.1
djangorestframework==3.11.0
gunicorn==20.0.4
Pillow==7.2.0
pytz==2020.1
sqlparse==0.3.1
whitenoise==5.2.0
wincertstore==0.2
